docker/dataprocessing/requirements.txt
kafka-python==2.0.2
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pathtools==0.1.2
Pillow==9.3.0
pipenv==2023.2.4
platformdirs==3.0.0
pluggy==1.0.0
PySocks==1.7.1
SQLAlchemy==2.0.7
text-unidecode==1.3
trio-websocket==0.10.2
typing_extensions==4.4.0
urllib3==1.26.15
psycopg2-binary==2.9.5
python-dateutil
